+++ 3d_detection_and_tracking/include/3d_detection_and_tracking/Type.h
@@ -40,7 +40,7 @@ struct Data
 
 struct Attributes
 {
-  uint state;
+  uint8_t state;
   double loc_x, loc_y, loc_z;
   double height, width, length;
   double rot_y;
